requirements.txt
tensorflow
matplotlib
numpy
pillow
scikit-learn
jupyter
